Return an empty Vec if available_ports() fails
Previously, we returned a Result from rfe::signal_generators() and
rfe::spectrum_analyzers() in order to pass along any errors from
serialport::available_ports(). Passing along such an uncommon error
isn't worth the extra effort of dealing with a Result type. So now, if
serialport::available_ports() returns an error, we consider it to be the
same thing as if no RF Explorers were found.