Check for truncated `Sweep`s
Sometimes, we receive `Sweep`s that are truncated and end with the
Early-End-of-Transmission (EEOT) byte sequence and/or a `Config` or
`SetupInfo` message. In order to make sure we don't miss messages at the
end of truncated `Sweep`s, we discard the bytes belonging to the `Sweep`
and attempt to parse a message from any remaining bytes.